Search Page Code Consolidation
-Moved out formatting and extraneous logic that doesn't use state values into a different file.
